upscale/requirements.txt
numpy==1.26.4
opencv-contrib-python==4.10.0.84
flask
celery
redis
pillow
opencv-python-headless
gunicorn
